+++ packages/cli/compiler/scripts/create-scripts-output-path.d.ts
@@ -1,0 +1,2 @@
+export declare const createScriptsOutputPath: () => string;
+//# sourceMappingURL=create-scripts-output-path.d.ts.map